requirements.txt
python-docx
google-cloud-storage
python_http_client
google-api-python-client
google-cloud-language

